Mirgrará o endpoint like de timeline para posts
A rota /timeline/post/:id/like para a ser agora /posts/:id/like
