+++ controllers/postsController.js
@@ -1,4 +1,5 @@
 import { postDeleter, postUpdate } from "../repositories/postsRepository.js";
+import { likeRepository, unlinkeRepository } from "../repositories/timelineRepositories.js";
 import { getUserByPostId } from "./../repositories/postsRepository.js";
 
 export async function deletePost(req, res) {
@@ -35,3 +36,27 @@ export async function updatePost(req, res) {
         res.sendStatus(500);
     };
 }
+
+export async function handleLike(req, res) {
+    try {
+        const {
+            userId,
+            postId,
+            liked,
+            likesPost
+        } = res.locals.likesData;
+  
+      if(liked) {
+        await unlinkeRepository(userId, postId);
+        
+      } else {
+        await likeRepository(userId, postId);
+      }
+
+      res.status(200).send(likesPost);
+  
+    } catch (e) {
+      console.log("Error in handleLike", e);
+      res.sendStatus(500);    
+    }
+  }
