+++ routes/postsRouter.js
@@ -4,7 +4,8 @@ import { validateToken } from "../middlewares/authMiddleware.js";
 import { validateSchema } from "../middlewares/schemaValidator.js";
 import  editSchema  from "../schemas/editSchema.js";
 
-import { deletePost, updatePost } from "../controllers/postsController.js";
+import { deletePost, handleLike, updatePost } from "../controllers/postsController.js";
+import { handleLikeMiddleware } from "../middlewares/postsMiddleware.js";
 
 const postsRouter = Router();
 
@@ -12,5 +13,6 @@ postsRouter.use(validateToken);
 
 postsRouter.delete('/posts/:postId', deletePost);
 postsRouter.put('/posts/:postId', validateSchema(editSchema), updatePost);
+postsRouter.post("/posts/:id/like", handleLikeMiddleware, handleLike);
 
 export default postsRouter;
